apiErrorHandling feat: integrate useFormErrorsHandler into useForgotPassword
