+++ features/forgotPassword/ui/ForgotPasswordForm/ForgotPasswordForm.tsx
@@ -1,97 +1,25 @@
 'use client'
-import { useState } from 'react'
-import { useGoogleReCaptcha } from 'react-google-recaptcha-v3'
-import { useForm } from 'react-hook-form'
 
-import { formWithEmailSchema } from '@/entities'
-import {
-  ControlledInput,
-  ErrorsMessage,
-  ErrorsMessagesResponse,
-  FormContainer,
-  SentEmailModal,
-  handleNetworkError,
-  handleServerError,
-  useAppDispatch,
-} from '@/shared'
-import { zodResolver } from '@hookform/resolvers/zod'
+import { ControlledInput, FormContainer, SentEmailModal } from '@/shared'
 import { Button, Recaptcha } from '@rambo-react/ui-meteors'
 import Link from 'next/link'
 
 import s from './ForgotPasswordForm.module.scss'
 
-import { useResetPasswordMutation } from '../../api'
+import { useForgotPassword } from '../../hooks'
 
 export function ForgotPasswordForm() {
-  const dispatch = useAppDispatch()
-  const [showModal, setShowModal] = useState(false)
-  const [recaptchaValue, setRecaptchaValue] = useState('')
-  const [sendLinkStatus, setSendLinkStatus] = useState<'error' | 'initial' | 'success'>('initial')
-  const [reCaptchaStatus, setReCaptchaStatus] = useState<
-    'checked' | 'expired' | 'initial' | 'loading' | 'withError'
-  >('initial')
-
-  const { control, handleSubmit, reset, setError, setValue, watch } = useForm({
-    defaultValues: { email: '' },
-    resolver: zodResolver(formWithEmailSchema),
-  })
-  const [validEmail, { isLoading: isLoadingValidEmail }] = useResetPasswordMutation()
-  const { executeRecaptcha } = useGoogleReCaptcha()
-
-  const handleCloseShowModal = () => {
-    setShowModal(false)
-    reset()
-  }
-
-  const handleSubmitDataForm = handleSubmit(async data => {
-    const response = await validEmail({ email: data.email, recaptchaValue })
-
-    if (response.error) {
-      if ('status' in response.error && response.error.status === 500) {
-        handleServerError(dispatch)
-      }
-      if ('status' in response.error && response.error.status === 'FETCH_ERROR') {
-        handleNetworkError(dispatch)
-      }
-      if ('data' in response.error) {
-        const errorReCaptcha = (response.error.data as ErrorsMessage).message
-
-        if (errorReCaptcha === 'Invalid reCAPTCHA') {
-          setSendLinkStatus('error')
-          setReCaptchaStatus('expired')
-        }
-      }
-      if ('data' in response.error) {
-        const errorMessage = (response.error.data as ErrorsMessagesResponse).errorsMessages[0]
-          .message
-
-        if (errorMessage === 'user not found') {
-          setError('email', { message: "User with this email doesn't exist", type: 'manual' })
-          setReCaptchaStatus('withError')
-        }
-      }
-    } else {
-      setSendLinkStatus('success')
-      setValue('email', data.email)
-      setShowModal(true)
-    }
-  })
-  const handleReCaptcha = async () => {
-    if (!executeRecaptcha) {
-      setError('email', { message: 'Execute recaptcha not yet available', type: 'manual' })
-
-      return
-    }
-
-    setReCaptchaStatus('loading')
-    const token = await executeRecaptcha('password_recovery')
-
-    setRecaptchaValue(token)
-    setReCaptchaStatus('checked')
-  }
-
-  const email = watch('email')
-  const isDisabled = !email || reCaptchaStatus !== 'checked' || isLoadingValidEmail
+  const {
+    control,
+    email,
+    handleCloseShowModal,
+    handleReCaptcha,
+    handleSubmitDataForm,
+    isDisabled,
+    reCaptchaStatus,
+    sendLinkStatus,
+    showModal,
+  } = useForgotPassword()
 
   return (
     <FormContainer title={'Forgot Password'}>
